Slice9 builder + runtime + panel widget + transform stack (#189) (#191)

+++ engine/ui/nt_ui_label.h
@@ -14,14 +14,13 @@ typedef struct nt_ui_context nt_ui_context_t;
 
 typedef struct {
     uint16_t font_id;         /* index into ctx->fonts[]; asserted < NT_UI_MAX_FONTS */
-    uint16_t font_size;       /* px; asserted > 0 */
+    float font_size;          /* px; asserted > 0. Passed to Clay as (uint16_t) for layout, float for rendering. */
     Clay_Color color;         /* 4 floats, 0..255 (Clay convention) */
     uint16_t line_height;     /* px; 0 = auto from font metrics */
     uint16_t letter_tracking; /* px; maps to Clay letterSpacing */
     uint8_t wrap_mode;        /* Clay_TextElementConfigWrapMode value; 0 = CLAY_TEXT_WRAP_WORDS default */
     uint8_t align;            /* Clay_TextAlignment value; 0 = CLAY_TEXT_ALIGN_LEFT default */
 } nt_ui_label_style_t;
-/* Layout: 2+2+16+2+2+1+1 = 26 B → padded to 32 B (alignof Clay_Color = 4). */
 _Static_assert(sizeof(nt_ui_label_style_t) <= 32, "nt_ui_label_style_t fits in 32 B");
 
 /* Asserts at entry: ctx, style, text non-NULL; font_id < NT_UI_MAX_FONTS;
@@ -32,6 +31,6 @@ _Static_assert(sizeof(nt_ui_label_style_t) <= 32, "nt_ui_label_style_t fits in 3
 void nt_ui_label(nt_ui_context_t *ctx, const nt_ui_element_data_t *data, const char *text, const nt_ui_label_style_t *style);
 
 /* nt_ui_label with font_size override (keeps style static-const; pair with nt_ui_fit_*). */
-void nt_ui_label_sized(nt_ui_context_t *ctx, const nt_ui_element_data_t *data, const char *text, const nt_ui_label_style_t *style, uint16_t font_size_override);
+void nt_ui_label_sized(nt_ui_context_t *ctx, const nt_ui_element_data_t *data, const char *text, const nt_ui_label_style_t *style, float font_size_override);
 
 #endif /* NT_UI_LABEL_H */
